# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "u-cafe",
        "description": "Documentation for u-cafe admin dashboard and main site",
        "contact": {
            "email": "s.murodjonov@uzinfocom.uz"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get list of Order",
                "description": "Returns list of Order",
                "operationId": "indexOrder",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Orders"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Insert a new Order",
                "description": "Insert a new Order",
                "operationId": "storeOrder",
                "requestBody": {
                    "description": "Order to create",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "data": {
                                        "$ref": "#/components/schemas/Order"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Order"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation exception"
                    }
                }
            }
        },
        "/api/orders/{order_id}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Show a Order from his Id",
                "description": "Show a Order from his Id",
                "operationId": "showOrder",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Order--id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Order"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Order not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Orders"
                ],
                "summary": "Delete a Order",
                "description": "Delete a Order",
                "operationId": "destroyOrder",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Order--id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Order"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Orders"
                ],
                "summary": "Update an existing Order",
                "description": "Update an existing Order",
                "operationId": "updateOrder",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Order--id"
                    }
                ],
                "requestBody": {
                    "description": "Order to update",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "data": {
                                        "$ref": "#/components/schemas/Order"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Order"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Order not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Order": {
                "title": "Order",
                "description": "Order model",
                "required": [],
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "id of Order",
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "name": {
                        "title": "name",
                        "description": "name of Order",
                        "type": "string",
                        "example": "Macbook Pro"
                    },
                    "sku": {
                        "title": "sku",
                        "description": "sku of Order",
                        "type": "string",
                        "example": "MCBPRO2022"
                    },
                    "price": {
                        "title": "price",
                        "description": "price of Order",
                        "type": "integer",
                        "example": "99"
                    },
                    "created_at": {
                        "title": "created_at",
                        "type": "dateTime",
                        "readOnly": "true",
                        "example": "2022-07-04T02:41:42.336Z"
                    },
                    "updated_at": {
                        "title": "updated_at",
                        "type": "dateTime",
                        "readOnly": "true",
                        "example": "2022-07-04T02:41:42.336Z"
                    }
                },
                "type": "object"
            },
            "Orders": {
                "title": "Orders list",
                "properties": {
                    "data": {
                        "title": "data",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Order"
                        }
                    },
                    "first_page_url": {
                        "title": "first_page_url",
                        "type": "string",
                        "example": "http://localhost:8080/api/merchant-customers?page=1"
                    },
                    "last_page_url": {
                        "title": "last_page_url",
                        "type": "string",
                        "example": "http://localhost:8080/api/merchant-customers?page=3"
                    },
                    "prev_page_url": {
                        "title": "prev_page_url",
                        "type": "string",
                        "example": "null"
                    },
                    "next_page_url": {
                        "title": "next_page_url",
                        "type": "string",
                        "example": "http://localhost:8080/api/merchant-customers?page=2"
                    },
                    "current_page": {
                        "title": "current_page",
                        "type": "integer",
                        "example": "1"
                    },
                    "from": {
                        "title": "from",
                        "type": "integer",
                        "example": "1"
                    },
                    "last_page": {
                        "title": "last_page",
                        "type": "integer",
                        "example": "3"
                    },
                    "path": {
                        "title": "path",
                        "type": "string",
                        "example": "http://localhost:8080/api/merchant-customers"
                    },
                    "per_page": {
                        "title": "per_page",
                        "type": "integer",
                        "example": "1"
                    },
                    "total": {
                        "title": "total",
                        "type": "integer",
                        "example": "3"
                    },
                    "links": {
                        "title": "links",
                        "type": "array",
                        "items": {
                            "properties": {
                                "url": {
                                    "title": "url",
                                    "type": "string",
                                    "example": "http://localhost:8080/api/merchant-customers?page=2"
                                },
                                "label": {
                                    "title": "label",
                                    "type": "string",
                                    "example": "1"
                                },
                                "active": {
                                    "title": "active",
                                    "type": "boolean",
                                    "example": "true"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "parameters": {
...
            "Order--id": {
                "name": "Order_id",
                "in": "path",
                "description": "Id of Order",
                "required": true,
                "schema": {
                    "type": "integer",
                    "example": "1"
                }
            },
...
        }
    }
}